+++ rambo/app.py
@@ -1,4 +1,3 @@
-import distutils
 import errno
 import os
 import platform
@@ -63,42 +62,65 @@ def _invoke_vagrant(cmd=None):
         os.close(fd)
     return p.returncode
 
+def set_init_vars(cwd=None, tmpdir=None):
+    """Set env vars that are used throughout Ruby and Python. If Ruby weren't used
+    this would be managed very differently, as these are effectively project scoped
+    global variables.
+
+    These calls are split into separate functions because on occaision we need to call
+    them more carefully than with this function, as in createproject.
+    """
+    set_env()
+    set_cwd(cwd)
+    set_tmpdir(tmpdir)
+
+
+def set_env():
+    set_env_var('ENV', PROJECT_LOCATION) # installed location of this code
+
+
 ## Defs used by main cli cmd
-def set_init_vars(cwd=None, tmpdir_path=None):
-    '''Set custom environment variables that are always going to be needed by
-    our custom Ruby code in the Vagrantfile chain.
+def set_cwd(cwd=None):
+    '''Set cwd environment variable and change the actual cwd to it.
 
     Args:
         cwd (path): Location of project (conf file, provisioning scripts, etc.).
-        tmpdir_path (path): Location of project's tmp dir.
     '''
-    # env vars available to Python and Ruby
-    set_env_var('ENV', PROJECT_LOCATION) # installed location of this code
-
     # effective CWD (likely real CWD, but may be changed by user.
     if cwd: # cli / api
-        set_env_var('CWD', cwd)
-    elif not get_env_var('CWD'): # Not previously set env var either
-        try:
-            set_env_var('CWD', os.getcwd())
-        except FileNotFoundError:
-            utils.abort('Your current working directory no longer exists. '
-                  'Did you delete it? Check for it with `ls ..`')
-    os.chdir(get_env_var('cwd'))
+        set_env_var('cwd', cwd)
+    elif get_env_var('cwd'):
+        pass # Already set - keep it.
+    else:
+        found_project = utils.find_conf(os.getcwd())
+        if found_project:
+            set_env_var('cwd', found_project)
+        else:
+            set_env_var('cwd', os.getcwd())
 
-    # loc of tmpdir_path
-    if tmpdir_path: # cli / api
-        set_env_var('TMPDIR_PATH',
-                    os.path.join(tmpdir_path, '.%s-tmp' % PROJECT_NAME))
-    elif get_env_var('TMPDIR_PATH'): # Previously set env var
-        set_env_var('TMPDIR_PATH',
-                    os.path.join(get_env_var('TMPDIR_PATH'),
+    os.chdir(get_env_var('cwd'))
+    return get_env_var('cwd')
+
+def set_tmpdir(tmpdir=None):
+    '''Set tmpdir and log_path locations. This should defaut to be inside the cwd.
+
+    Args:
+        tmpdir (path): Location of project's tmp dir.
+    '''
+    # loc of tmpdir
+    if tmpdir: # cli / api
+        set_env_var('TMPDIR',
+                    os.path.join(tmpdir, '.%s-tmp' % PROJECT_NAME))
+    elif get_env_var('TMPDIR'): # Previously set env var
+        set_env_var('TMPDIR',
+                    os.path.join(get_env_var('TMPDIR'),
                                  '.%s-tmp' % PROJECT_NAME))
     else: # Not set, set to default loc
-        set_env_var('TMPDIR_PATH',
+        set_env_var('TMPDIR',
                     os.path.join(os.getcwd(),
                                  '.%s-tmp' % PROJECT_NAME)) # default (cwd)
-    set_env_var('LOG_PATH', os.path.join(get_env_var('TMPDIR_PATH'), 'logs'))
+
+    set_env_var('LOG_PATH', os.path.join(get_env_var('TMPDIR'), 'logs'))
 
 def set_vagrant_vars(vagrant_cwd=None, vagrant_dotfile_path=None):
     '''Set the environment varialbes prefixed with `VAGRANT_` that vagrant
@@ -124,26 +146,32 @@ def set_vagrant_vars(vagrant_cwd=None, vagrant_dotfile_path=None):
         os.environ['VAGRANT_DOTFILE_PATH'] = os.path.normpath(os.path.join(os.getcwd(), '.vagrant')) # default (cwd)
 
 ## Defs for cli subcommands
-def createproject(project_name, config_only=None):
+def createproject(project_name, cwd, tmpdir, config_only=None, ctx=None):
     '''Create project with basic configuration files.
 
     Agrs:
         project_name (path): Place to create a new project. Must be non-existing dir.
         config_only (bool): Determins if we should only place a conf file in the new project.
     '''
-    ## Create project dir
-    path = os.path.join(os.getcwd(), project_name)
+    # initialize paths
+    set_env()
+    cwd = set_cwd(cwd)
+    path = os.path.join(cwd, project_name)
+    set_tmpdir(path)
+
+    # create new project dir
     try:
         os.makedirs(path) # Make parent dirs if needed.
     except FileExistsError:
         utils.abort('Directory already exists.')
     utils.echo('Created %s project "%s" in %s.'
                % (PROJECT_NAME.capitalize(), project_name, path))
-    ## Fill project dir with basic configs.
-    install_config(output_path=path)
+
+    # Fill project dir with basic configs.
+    install_config(ctx, output_path=path)
     if not config_only:
         export('saltstack', path)
-        install_auth(output_path=path)
+        install_auth(ctx, output_path=path)
 
 def destroy(ctx=None, **params):
     '''Destroy a VM / container and all its metadata. Default leaves logs.
@@ -159,15 +187,27 @@ def destroy(ctx=None, **params):
     # TODO add an --all flag to delete the whole .rambo-tmp dir. Default leaves logs.
 
     if not ctx: # Using API. Else handled by cli.
-        set_init_vars(params.get('cwd'), params.get('tmpdir_path'))
+        set_init_vars(params.get('cwd'), params.get('tmpdir'))
         set_vagrant_vars(params.get('vagrant_cwd'), params.get('vagrant_dotfile_path'))
 
-    vagrant_general_command('destroy --force')
-    utils.file_delete(os.path.join(get_env_var('TMPDIR_PATH'), 'provider'))
-    utils.file_delete(os.path.join(get_env_var('TMPDIR_PATH'), 'random_tag'))
-    utils.dir_delete(os.environ.get('VAGRANT_DOTFILE_PATH'))
-    utils.echo('Temporary files removed')
-    utils.echo('Destroy complete.')
+    destroy_cmd = vagrant_general_command('destroy --force')
+
+    # If there's any error code from Vagrant, don't delete the metadata.
+    if not destroy_cmd:  # I.e. we succeeded - ret code == 0
+        utils.file_delete(os.path.join(get_env_var('TMPDIR'), 'provider'))
+        utils.file_delete(os.path.join(get_env_var('TMPDIR'), 'random_tag'))
+        utils.dir_delete(os.environ.get('VAGRANT_DOTFILE_PATH'))
+        utils.echo('Temporary files removed')
+
+        if params.get("vm_name"):  # Additionally remove the box if we can.
+            utils.echo(f"Now removing base VirtualBox data for VM {params['vm_name']}.")
+            os.system(f"vboxmanage controlvm {params['vm_name']} poweroff")
+            os.system(f"vboxmanage unregistervm {params['vm_name']} --delete")
+
+        utils.echo('Destroy complete.')
+    else:
+        utils.echo('We received an error. Destroy may not be complete.')
+
 
 def export(resource=None, export_path=None, force=None):
     '''Drop default code in the CWD / user defined space. Operate on saltstack
@@ -218,14 +258,14 @@ def export(resource=None, export_path=None, force=None):
 
 def halt(ctx=None, *args, **params):
     if not ctx: # Using API. Else handled by cli.
-        set_init_vars(params.get('cwd'), params.get('tmpdir_path'))
+        set_init_vars(params.get('cwd'), params.get('tmpdir'))
         set_vagrant_vars(params.get('vagrant_cwd'), params.get('vagrant_dotfile_path'))
     else:
         args = ctx.args + list(args)
 
     vagrant_general_command('{} {}'.format('halt', ' '.join(args)))
 
-def install_auth(ctx=None, output_path=None, **params):
+def install_auth(ctx=None, output_path=None, **kwargs):
     '''Install auth directory.
 
     Agrs:
@@ -233,7 +273,7 @@ def install_auth(ctx=None, output_path=None, **params):
         output_path (path): Path to place auth dir.
     '''
     if not ctx: # Using API. Else handled by cli.
-        set_init_vars(params.get('cwd'), params.get('tmpdir_path'))
+        set_init_vars(params.get('cwd'), params.get('tmpdir'))
 
     if not output_path:
         output_path = get_env_var('cwd')
@@ -261,7 +301,7 @@ def install_auth(ctx=None, output_path=None, **params):
     # and not needing the auth key scripts.
     # load_provider_keys()
 
-def install_config(ctx=None, output_path=None, **params):
+def install_config(ctx=None, output_path=None, **kwargs):
     '''Install config file.
 
     Agrs:
@@ -269,7 +309,7 @@ def install_config(ctx=None, output_path=None, **params):
         output_path (path): Path to place conf file.
     '''
     if not ctx: # Using API. Else handled by cli.
-        set_init_vars(params.get('cwd'), params.get('tmpdir_path'))
+        set_init_vars(params.get('cwd'), params.get('tmpdir'))
 
     if not output_path:
         output_path = get_env_var('cwd')
@@ -309,7 +349,7 @@ def scp(ctx=None, locations=None, **params):
     which allows for simplified args.
     '''
     if not ctx: # Using API. Else handled by cli.
-        set_init_vars(params.get('cwd'), params.get('tmpdir_path'))
+        set_init_vars(params.get('cwd'), params.get('tmpdir'))
         set_vagrant_vars(params.get('vagrant_cwd'), params.get('vagrant_dotfile_path'))
 
     if len(locations)!=2:
@@ -339,7 +379,7 @@ def ssh(ctx=None, command=None, **params):
         vagrant_dotfile_path (path): Location of `.vagrant` metadata directory. Used if invoked with API only.
     '''
     if not ctx: # Using API. Else handled by cli.
-        set_init_vars(params.get('cwd'), params.get('tmpdir_path'))
+        set_init_vars(params.get('cwd'), params.get('tmpdir'))
         set_vagrant_vars(params.get('vagrant_cwd'), params.get('vagrant_dotfile_path'))
 
     ## Add pass-through 'command' option.
@@ -361,30 +401,50 @@ def up(ctx=None, **params):
     In params, this looks for:
         provider (str): Provider to use.
         box (str): Vagrant box to use.
+        cpus (int): Number of CPUs to give VirtualBox VM.
         guest_os (str): Guest OS to use.
         ram_size (int): RAM in MB to use.
         drive_size (int): Drive size in GB to use.
         machine_type (str): Machine type to use for cloud providers.
-        sync_dir (path): Path to sync into VM.
+        sync_dirs (path): Paths to sync into VM.
+        sync_type (str): Type of syncing to use.
+        ports (str): Ports to forward.
         provision (bool): vagrant provisioning flag.
+        provision_cmd (str): Command used at beginning of provisioning.
+        provision_script (path): Path to script to use for provisioning.
+        provision_with_salt (bool): Flag to indicate provisioning with salt.
+        provision_with_salt_legacy (bool): Flag to indicate provisioning with salt using the legacy style.
         destroy_on_error (bool): vagrant destroy-on-error flag.
         vagrant_cwd (path): Location of `Vagrantfile`. Used if invoked with API only.
         vagrant_dotfile_path (path): Location of `.vagrant` metadata directory. Used if invoked with API only.
+        vm_name (str): Name of the VM or container.
     '''
     # TODO: Add registering of VM for all of this installation to see
-
     if not ctx: # Using API. Else handled by cli.
-        set_init_vars(params.get('cwd'), params.get('tmpdir_path'))
+        set_init_vars(params.get('cwd'), params.get('tmpdir'))
         set_vagrant_vars(params.get('vagrant_cwd'), params.get('vagrant_dotfile_path'))
 
-    ## Option Handling - These might modify the params dict or set env vars.
+    ## Option Handling - These might modify the params dict and/or set env vars.
     params['guest_os'] = options.guest_os_option(params.get('guest_os'))
     params['box'] = options.box_option(params.get('box'))
+    params['cpus'] = options.cpus_option(params.get('cpus'))
+    params['hostname'] = options.hostname_option(params.get('hostname'))
     params['machine_type'] = options.machine_type_option(params.get('machine_type'), params.get('provider'))
+    params['project_dir'] = options.project_dir_option(params.get('project_dir'))
     params['provider'] = options.provider_option(params.get('provider'))
+    params['provision_cmd'] = options.provision_cmd_option(params.get('provision_cmd'))
+    params['provision_script'] = options.provision_script_option(params.get('provision_script'))
+    params['provision_with_salt'] = options.provision_with_salt_option(params.get('provision_with_salt'))
+    params['provision_with_salt_legacy'] = options.provision_with_salt_legacy_option(params.get('provision_with_salt_legacy'))
     params['ram_size'], params['drive_size'] = options.size_option(
         params.get('ram_size'), params.get('drive_size')) # both ram and drive size
-    params['sync_dir'] = options.sync_dir_option(params.get('sync_dir'))
+    params['salt_bootstrap_args'] = options.salt_bootstrap_args_option(params.get('salt_bootstrap_args'))
+    params['sync_dirs'] = options.sync_dirs_option(params.get('sync_dirs'))
+    params['sync_type'] = options.sync_type_option(params.get('sync_type'))
+    params['ports'] = options.ports_option(params.get('ports'))
+    params['vm_name'] = options.vm_name_option(params.get('vm_name'))
+
+    cmd = 'up'
 
     ## Provider specific handling.
     ## Must come after all else, because logic may be done on params above.
@@ -393,10 +453,11 @@ def up(ctx=None, **params):
     elif params['provider'] == 'docker':
         vagrant_providers.docker()
     elif params['provider'] == 'ec2':
-        vagrant_providers.ec2()
+        vagrant_providers.ec2(**params)
+    else:
+        cmd += " --provider={}".format(params['provider'])
 
     ## Add straight pass-through flags. Keep test for True/False explicit as only those values should work
-    cmd = 'up'
     if params.get('provision') is True:
         cmd = '{} {}'.format(cmd, '--provision')
     elif params.get('provision') is False:
@@ -416,7 +477,7 @@ def vagrant_general_command(cmd):
         cmd (str): String to append to command `vagrant ...`
     '''
     # Modify cmd in private function to keep enforcement of being a vagrant cmd there.
-    _invoke_vagrant(cmd)
+    return _invoke_vagrant(cmd)
 
 
 class Run_app():