+++ rambo/vagrant/modules.rb
@@ -2,14 +2,16 @@ require 'securerandom'
 
 def random_tag
   host = `hostname`.strip # Get the hostname from the shell and removing trailing \n
-  tmp_dir = get_env_var_rb('TMPDIR_PATH') || File.join(Dir.pwd, '.' + PROJECT_NAME + '-tmp')
+  tmp_dir = get_env_var_rb('TMPDIR') || File.join(Dir.pwd, '.' + PROJECT_NAME + '-tmp')
   Dir.mkdir(tmp_dir) unless Dir.exist?(tmp_dir)
   random_tag_path = File.join(tmp_dir, 'random_tag')
   if File.file?(random_tag_path)
     tag = File.read(random_tag_path)
   else
     # 95 is unlikely to be used. It is the ascii code for an underscore
-    guest_hostname = host + '-' + File.basename(File.dirname(tmp_dir)).sub("_", "95")
+    replacements = {'_' => '95'}
+    guest_hostname = host + '-' + File.basename(File.dirname(tmp_dir))
+    guest_hostname = guest_hostname.gsub(Regexp.union(replacements.keys), replacements)
     guest_hostname = truncate(guest_hostname, 43) # 64 - 15 for the next line - 6 for "rambo-"
     tag = guest_hostname + '-' + SecureRandom.hex(6)
     File.write(random_tag_path, tag)
@@ -18,7 +20,7 @@ def random_tag
 end
 
 def read_provider_file
-  tmp_dir = get_env_var_rb('TMPDIR_PATH') || File.join(Dir.pwd, '.' + PROJECT_NAME + '-tmp')
+  tmp_dir = get_env_var_rb('TMPDIR') || File.join(Dir.pwd, '.' + PROJECT_NAME + '-tmp')
   provider_path = File.join(tmp_dir, 'provider')
   if File.file?(provider_path)
     provider=''
@@ -32,7 +34,7 @@ def read_provider_file
 end
 
 def write_provider_file(provider)
-  tmp_dir = get_env_var_rb('TMPDIR_PATH') || File.join(Dir.pwd,  '.' + PROJECT_NAME + '-tmp')
+  tmp_dir = get_env_var_rb('TMPDIR') || File.join(Dir.pwd,  '.' + PROJECT_NAME + '-tmp')
   provider_path = File.join(tmp_dir, 'provider')
   File.write(provider_path, provider)
 end